--- src/components/styled/GradientBackgroundView.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-import LinearGradient from 'react-native-linear-gradient';
-
-export default styled(LinearGradient).attrs({
-    colors: ['#5DD39E', '#348EAC'],
-})`
-    flex: 1;
-`;
