--- src/components/styled/SnowSpinner.js
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-import Spinner from 'react-native-spinkit';
-import { scale } from '../../utils/scallingUtils';
-
-export default styled(Spinner).attrs({
-    type: "ThreeBounce",
-    color: "#ffffff",
-    size: scale(72)
-})`
-    position: absolute;
-    bottom: 10%;
-`;
